Extract ensureSidePanelClosed
