+++ packages/e2e-tests/tests/stacking.test.ts
@@ -8,7 +8,10 @@ import {
   selectElementsRowWithText,
 } from "../helpers/elements-panel";
 import { toggleToolboxLayout } from "../helpers/layout";
-import { getBreakpointsAccordionPane } from "../helpers/pause-information-panel";
+import {
+  ensureSidePanelClosed,
+  getBreakpointsAccordionPane,
+} from "../helpers/pause-information-panel";
 import { mapLocators, waitFor } from "../helpers/utils";
 import test from "../testFixtureCloneRecording";
 
@@ -143,18 +146,6 @@ const testCases: StackingTestCase[] = [
     expectedRules: ["element", "body > div"],
   },
 ];
-
-async function ensureSidePanelClosed(page: Page) {
-  // Clicks that aren't directly on an element can cause the "Comments" pane to open.
-  // Ensure that it's closed by forcing the "Pause" pane to open instead...
-  const pane = getBreakpointsAccordionPane(page);
-  const pauseButton = page.locator('[data-test-name="ToolbarButton-PauseInformation"]');
-  await pauseButton.click();
-  const isVisible = await pane.isVisible();
-  if (isVisible) {
-    await pauseButton.click();
-  }
-}
 
 async function verifySelectedElementUnderCursor(
   page: Page,
